requirements.txt
gym==0.26.2
pygame==2.5.2
hydra-core==1.3.2
moviepy==1.0.3
scikit-learn==1.3.1
seaborn==0.13.0
torch>=2.1.0
torchvision>=0.16.0
tensordict>=0.2.1
torchrl>=0.2.1
tqdm==4.66.1
wandb==0.15.12
causal-learn>=0.1.3.7
FrEIA>=0.2